Breaks down the different components into organized files
