+++ src/schedulers/slot.scheduler.ts
@@ -1,0 +1,9 @@
+import { type SlotSchedule } from './utils/slot-schedule'
+
+export abstract class SlotScheduler<T> {
+  public abstract generateSchedule(
+    n: number,
+    currentSlot: number,
+    input: T
+  ): SlotSchedule
+}
