+++ src/schedulers/utils/time-input.ts
@@ -1,0 +1,7 @@
+export class TimeInput {
+  distance: number
+
+  constructor(distance: number) {
+    this.distance = distance
+  }
+}
